+++ docs/webpack.config.babel.js
@@ -52,16 +52,21 @@ export const makeConfig = (config = {}) => {
           test: /\.css$/,
           loader: ExtractTextPlugin.extract(
             "style-loader",
-            "css-loader" + (
-              "?modules"+
-              "&localIdentName=" +
-              (
-                process.env.NODE_ENV === "production"
+            [ `css-loader?modules&localIdentName=${
+                config.production
                 ? "[hash:base64:5]"
                 : "[path][name]--[local]--[hash:base64:5]"
-              ).toString()
-            ) + "!" +
-            "postcss-loader",
+              }`,
+              "postcss-loader",
+            ].join("!"),
+          ),
+          exclude: /\.global\.css$/,
+        },
+        {
+          test: /\.global\.css$/,
+          loader: ExtractTextPlugin.extract(
+            "style-loader",
+            [ "css-loader", "postcss-loader" ].join("!"),
           ),
         },
         {